+++ draft-packages/stories/Select.stories.tsx
@@ -1,13 +1,18 @@
+import * as colorTokens from "@kaizen/design-tokens/tokens/color.json"
 import { AsyncSelect, Select } from "@kaizen/draft-select"
 import * as React from "react"
 
-const StoryContainer = ({ children }: { children: React.ReactNode }) => {
-  return <div style={{ width: "300px", margin: "12px auto" }}>{children}</div>
-}
+const StoryContainer = ({ children }: { children: React.ReactNode }) => (
+  <div style={{ width: "300px", margin: "12px auto" }}>{children}</div>
+)
 
-const WideStoryContainer = ({ children }: { children: React.ReactNode }) => {
-  return <div style={{ width: "500px", margin: "12px auto" }}>{children}</div>
-}
+const WideStoryContainer = ({ children }: { children: React.ReactNode }) => (
+  <div style={{ width: "500px", margin: "12px auto" }}>{children}</div>
+)
+
+const NarrowStoryContainer = ({ children }: { children: React.ReactNode }) => (
+  <div style={{ width: "200px", margin: "12px auto" }}>{children}</div>
+)
 
 const options = [
   { value: "Mindy", label: "Mindy" },
@@ -42,11 +47,10 @@ const asyncOptions = [
   { value: "Roberto", label: "Roberto" },
 ]
 
-const filterNames = (inputValue: string) => {
-  return asyncOptions.filter(i =>
+const filterNames = (inputValue: string) =>
+  asyncOptions.filter(i =>
     i.label.toLowerCase().includes(inputValue.toLowerCase())
   )
-}
 
 const promiseOptions = inputValue =>
   new Promise(resolve => {
@@ -69,7 +73,49 @@ export default {
 
 export const Single = () => (
   <StoryContainer>
-    <Select options={options} placeholder="Placeholder" isSearchable={false} />
+    <Select
+      options={options}
+      placeholder="Placeholder"
+      isSearchable={false}
+      isDisabled={false}
+      defaultValue={options[0]}
+    />
+  </StoryContainer>
+)
+
+export const SingleEllipsis = () => {
+  const localOptions = [
+    {
+      value: "a",
+      label:
+        "Long option where the container is 200px fixed width and the selected option should ellipsize",
+    },
+    { value: "b", label: "B" },
+  ]
+
+  return (
+    <NarrowStoryContainer>
+      <Select
+        options={localOptions}
+        isSearchable={false}
+        defaultValue={localOptions[0]}
+      />
+    </NarrowStoryContainer>
+  )
+}
+
+SingleEllipsis.story = {
+  name: "Single with ellipsizing selection",
+}
+
+export const SingleClearable = () => (
+  <StoryContainer>
+    <Select
+      options={options}
+      placeholder="Placeholder"
+      isSearchable={false}
+      isClearable={true}
+    />
   </StoryContainer>
 )
 
@@ -109,7 +155,124 @@ export const MultiAsyncSearchable = () => (
     />
   </WideStoryContainer>
 )
-
 MultiAsyncSearchable.story = {
   name: "Multi-Async Searchable",
+}
+
+export const SingleSecondary = () => (
+  <StoryContainer>
+    <Select
+      options={options}
+      isSearchable={false}
+      defaultValue={options[0]}
+      variant="secondary"
+    />
+  </StoryContainer>
+)
+
+SingleSecondary.story = {
+  name: "Single, Secondary",
+}
+
+export const SingleSecondaryDisabled = () => (
+  <StoryContainer>
+    <Select
+      options={options}
+      isDisabled={true}
+      isSearchable={false}
+      defaultValue={options[0]}
+      variant="secondary"
+    />
+  </StoryContainer>
+)
+
+SingleSecondaryDisabled.story = {
+  name: "Single, Secondary, Disabled",
+}
+
+export const SingleSecondaryReversed = () => (
+  <StoryContainer>
+    <Select
+      options={options}
+      isSearchable={false}
+      defaultValue={options[0]}
+      variant="secondary"
+      reversed={true}
+    />
+  </StoryContainer>
+)
+
+SingleSecondaryReversed.story = {
+  name: "Single, Secondary, Reversed",
+  parameters: {
+    backgrounds: [
+      {
+        name: "Wisteria 700",
+        value: colorTokens.kz.color.wisteria[700],
+        default: true,
+      },
+    ],
+  },
+}
+
+export const SingleSecondaryWithEllipsis = () => {
+  const localOptions = [
+    {
+      value: "a",
+      label:
+        "Long option where the container is 200px fixed width and the selected option should ellipsize",
+    },
+    { value: "b", label: "B" },
+    { value: "c", label: "Regular" },
+  ]
+  return (
+    <NarrowStoryContainer>
+      <Select
+        options={localOptions}
+        isSearchable={false}
+        defaultValue={localOptions[0]}
+        variant="secondary"
+        reversed={true}
+      />
+    </NarrowStoryContainer>
+  )
+}
+
+SingleSecondaryWithEllipsis.story = {
+  name: "Single Secondary with ellipsis",
+  parameters: {
+    backgrounds: [
+      {
+        name: "Wisteria 700",
+        value: colorTokens.kz.color.wisteria[700],
+        default: true,
+      },
+    ],
+  },
+}
+
+export const SingleSecondaryReversedDisabled = () => (
+  <StoryContainer>
+    <Select
+      options={options}
+      isDisabled={true}
+      isSearchable={false}
+      defaultValue={options[0]}
+      variant="secondary"
+      reversed={true}
+    />
+  </StoryContainer>
+)
+
+SingleSecondaryReversedDisabled.story = {
+  name: "Single Secondary Reversed (disabled)",
+  parameters: {
+    backgrounds: [
+      {
+        name: "Wisteria 700",
+        value: colorTokens.kz.color.wisteria[700],
+        default: true,
+      },
+    ],
+  },
 }