+++ draft-packages/title-block-zen/KaizenDraft/TitleBlockZen/SecondaryActions.tsx
@@ -1,19 +1,7 @@
-import { Button, ButtonProps, IconButton } from "@kaizen/draft-button"
-import {
-  Menu,
-  MenuContent,
-  MenuItem,
-  MenuItemProps,
-  MenuProps,
-} from "@kaizen/draft-menu"
+import { Button, IconButton } from "@kaizen/draft-button"
+import { Menu, MenuContent, MenuItem } from "@kaizen/draft-menu"
 import * as React from "react"
-import {
-  ButtonWithOnClickOrHref,
-  isMenuGroupNotButton,
-  isMenuItemNotButton,
-  MenuGroup,
-  SecondaryActionsProps,
-} from "./TitleBlockZen"
+import { SecondaryActionsProps, TitleBlockMenuItemProps } from "./TitleBlockZen"
 import Toolbar from "./Toolbar"
 const chevronDownIcon = require("@kaizen/component-library/icons/chevron-down.icon.svg")
   .default
@@ -24,17 +12,18 @@ const styles = require("./TitleBlockZen.scss")
 
 type Props = {
   secondaryActions?: SecondaryActionsProps
-  secondaryOverflowMenuItems?: MenuItemProps[]
+  secondaryOverflowMenuItems?: TitleBlockMenuItemProps[]
   reversed?: boolean
 }
 
 const renderSecondaryOverflowMenu = (
-  secondaryOverflowMenuItems?: MenuItemProps[],
+  secondaryOverflowMenuItems?: TitleBlockMenuItemProps[],
   reversed?: boolean
 ) => {
-  if (!secondaryOverflowMenuItems) return null
+  if (!secondaryOverflowMenuItems) return undefined
   return (
     <Menu
+      align="right"
       button={<IconButton label="" reversed={reversed} icon={meatballsIcon} />}
     >
       <MenuContent>
@@ -51,42 +40,60 @@ const SecondaryActions = ({
   secondaryOverflowMenuItems,
   reversed = false,
 }: Props) => {
-  if (!secondaryActions) return null
+  if (!secondaryActions && !secondaryOverflowMenuItems) return null
 
-  const stuff = secondaryActions.map(a => {
-    if (isMenuGroupNotButton(a)) {
-      return (
-        <Menu
-          button={
+  const secondaryActionsAsToolbarItems = secondaryActions
+    ? secondaryActions.map(a => {
+        if ("menuItems" in a) {
+          return (
+            <Menu
+              align="right"
+              button={
+                <Button
+                  secondary
+                  label={a.label}
+                  reversed={reversed}
+                  icon={chevronDownIcon}
+                  iconPosition="end"
+                />
+              }
+            >
+              <MenuContent>
+                {a.menuItems.map(menuItem => (
+                  <MenuItem {...menuItem} />
+                ))}
+              </MenuContent>
+            </Menu>
+          )
+        } else {
+          if ("onClick" in a && "href" in a) {
+            // eslint-disable-next-line no-console
+            console.warn(
+              "\u001b[33m \nTITLE BLOCK WARNING:\nSecondary actions only support " +
+                "either an href or an onClick, not both simultaneously.\n"
+            )
+          }
+          return (
             <Button
               secondary
-              label={a.label}
               reversed={reversed}
-              icon={chevronDownIcon}
-              iconPosition="end"
+              {...a}
+              automationId="title-block-secondary-actions-button"
             />
-          }
-        >
-          <MenuContent>
-            {a.menuItems.map(menuItem => (
-              <MenuItem {...menuItem} />
-            ))}
-          </MenuContent>
-        </Menu>
-      )
-    } else {
-      return <Button secondary reversed={reversed} {...a} />
-    }
-  })
+          )
+        }
+      })
+    : []
 
-  let toolbarItems = stuff
   const overflowMenu = renderSecondaryOverflowMenu(
     secondaryOverflowMenuItems,
     reversed
   )
-  if (overflowMenu !== null) {
-    toolbarItems = [...toolbarItems, overflowMenu]
-  }
+
+  const toolbarItems = [
+    ...secondaryActionsAsToolbarItems,
+    ...(overflowMenu ? [overflowMenu] : []),
+  ]
 
   return (
     <div className={styles.secondaryActionsContainer}>