+++ draft-packages/zen-off-canvas/KaizenDraft/ZenOffCanvas/components/Menu.tsx
@@ -10,18 +10,16 @@ type Props = {
   link: MenuProps
 }
 
-const Menu = ({ link, section }: Props) => {
-  return (
-    <div
-      className={classNames(styles.menu, {
-        [styles.final]: section === "final",
-        [styles.active]: !!link.active,
-      })}
-      key={`${link.heading}-${uuidv4()}`}
-    >
-      {link}
-    </div>
-  )
-}
+const Menu = ({ link, section }: Props) => (
+  <div
+    className={classNames(styles.menu, {
+      [styles.final]: section === "final",
+      [styles.active]: !!link.active,
+    })}
+    key={`${link.heading}-${uuidv4()}`}
+  >
+    {link}
+  </div>
+)
 
 export default Menu